Reformat the source code
The source code style (brace style, tab vs space...) was inconsistent,
so it needed a clean up. I took this opportunity to change everything to
something I like.

Changing the entire file makes it hard to use git blame, but it was
already broken because of the large code movement made around 0.4.0.