Get rid of non working code
There was a bunch of cpuFreq code lying around, but the critical part
about getting the data didn't work. Maybe this should be fixed, but in
the meanwhile, there is no point in keeping the rest of the system.

Also, when resising history, the past would be filled with invalid data.
Zeroing it is both quicker and more accurate.