Fix the border
Put the border on the right component, and use the right size, to match
the visual style of other xfce pannel plugins. On the other hand, the
border takes space. Making it optional allows for more efficient use of
screen real estate if wanted. This resolves bug 6257 and the first issue
of bug 5433.