Autodetect the default command.
xfce4-taskmanager is a better default command than top, but I was using
top as a default because xfce4-taskmanager isn't guaranteed to be around.
With this change, xfce4-taskmanaer will be used as a default if
available, and top will be used instead if not. The autodetection is
only used when a there are no saved settings, that is, only when adding
a new cpugraph to the pannel.

Adapted from a patch by Mike Massonnet.