Improve the migration from the previous release
In the previous release (0.4.0), the associated command was neither
launched in a terminal nor using startup notifications. Now, depending
on whether xfce4-taskmanager is present in the system, one of these
settings will be turned on by default when creating a new cpugraph, or
when running this version of cpugraph for the first time on a system
that has a setting file for 0.4.0.

The default command of 0.4.0 was "xterm -e top". Running this with
either startup notification or in a terminal is not appropriate.

The name of the entry in the rc file used to save the associated command
is now different from the one used in 0.4.0. This means that the old
setting, if it exists, will be ignored, and cpugraph will use the new
default setting. For people who had kept the old default, people who had
changed it to xfce4-taskmanager, or who had changed it to run top in
their favorite terminal, this is likely to be the best thing to do, and
I suspect that these cases cover most people.

People who had set a different custom command will need to set it again,
but depending on what their actual command is, they probably would have
need to tweak it anyway, because of the new "Startup notification" and
"Run in terminal" options.