Implement 4819: support for tracking only one cpu
It is now possible to track only one cpu or core in a particular
cpugraph, even on systems that have many cpus or cores.

The os dependent part still fetches data for all cores, but that data is
only used to display information about one of them. This helps keep the
design simple.