Lower CPU usage when rendering bars
This patch lowers the CPU usage of the cpugraph plugin itself
by a factor of 3 when rendering bars on a CPU with 16 threads.
This is achieved by rendering the bars using Cairo instead of GTK.

The top pixel of a bar is alpha-blended with the background.
This can help the user to read CPU usage more accurately
on displays with a resolution of 1080p or smaller where
the height (or width, depending on panel orientation)
of the XFCE panel can be smaller than 32 pixels.

The width (or height, depending on panel orientation)
of the section displaying bars is smaller by about 25%,
which makes it better suited for modern many-core CPUs.
(With a different GTK style than the one used for testing,
it may be more than 25%.)

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/8
Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/6
See also: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/7