Use floating-point instead of fixed-point CPU load values
This patch removes CPU_SCALE=256 in favor of storing CPU load
and history values as 32-bit floats. A small benefit of
this change is that this removes truncation inaccuracies
resulting from the conversion of CPU load values usually
sampled in /proc/stat with resolution of 100 Hz into the
now-removed fixed-point resolution of 256 (8 bits).

The CPU history graph draw area is now displaying the last
point (or last line, depending on display mode) using
alpha-blending.

The thickness of the CPU usage line in grid mode has been
increased from 1.0 to 1.75 to make the line appear to be
above the grid rather than below the grid.