Dynamic default command lookup
The default command was being determined once (when plugin settings
were being loaded). This patch lets the plugin make the lookup
every time the user requests to run the plugin's "associated command".

Updated list of default commands (ordered by priority):

  1. xfce4-taskmanager
  2. htop (new with this patch)
  3. top

Making the lookup dynamic covers the following use case:

  - If the user installs or uninstalls xfce4-taskmanager
    then the cpugraph plugin does not need to be restarted
    and does not need to reconfigured by the user (assuming
    that the "Associated command" field is already empty)

Lastly, if the "Associated command" field in the configuration dialog
is empty, then the checkboxes below this field are now disabled in order
to signal to the user that they have no effect.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/5