Clamp CPU usage values to zero if they are below a threshold
Values below the threshold are clamped to zero when painting
the CPU usage graph and CPU usage bars. This can be used to decrease
the amount of unwanted visual distractions caused by the cpugraph plugin.

If this feature is combined with a transparent background color
and with disabled option "Show frames", the cpugraph plugin
becomes effectively invisible unless the threshold is exceeded.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/2