Scale CPU load history chart according to the current update interval
When displaying the CPU load history chart, the algorithm determines
the CPU load values which are stored in the so-far recorded history nearest
to the wall-clock timestamps of the values to be displayed in the chart.

The capacity of the circular history buffer has been increased
from 128 items to 128*slowest/fastest = 1536 items.

This enables more seamless/accurate switching between
the Slowest-Slow-Normal-Fast-Fastest update intervals in cpugraph's
properties dialog.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/17