Change semantics of the non-linear time-scale mode
This patch changes the semantics of the non-linear mode from scaling the
y-axis (CPU load) to scaling the x-axis (time). This enables the cpugraph
plugin to display CPU load history up to 16 hours into the past using
a width of just 128 pixels in case the update interval is 3 seconds.

 Width:               128  60   128   60
 Interval:            3    3    0.25  0.25
 => Linear range:     6m   3m   32s   15s
 => Non-linear range: 16h  31m  80m   2.6m

The implementation uses an exponential function with base 1.04.

The behavior of scaling the y-axis (CPU load) according to how far it is
in the past can be restored in a future patch upon request.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/18
Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/19
Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/20