Add support for per-core history graphs
This patch enables the plugin to display the histories of all CPU cores
next to each other.

The amount of spacing between the history graphs currently depends on
the "Show border" setting: if "Show border" is enabled then the spacing
between cores is 1 pixel, otherwise it is zero.

For core counts higher than 4, this can make the plugin take a lot of space
in the panel (depending on the size of displayed history).

The implementation internally stores the histories of all CPU cores
irrespective of whether "Per-core history graphs" setting is enabled or
disabled. This has almost no negative performance impact, but it can
increase the plugin's memory consumption on many-core CPUs if the history
size is large.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/15