Enable user-defined spacing between per-core history graphs
The default value of the spacing is 1 pixel, maximum is 3 pixels. The
maximum spacing might need to be increased, or computed in a different way,
when 8K displays with DPI 300 become more common. The color of the spacing
is the Xfce panel background color.

See also: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/15