Add xfce4::RGBA
An RGB color with an alpha channel. All color channels are 64-bit floats.

The operators == and != are unsupported. The RGBA::equals() member function
takes an epsilon (defaults to 1e-10 if omitted) which is used as a maximum
allowed difference between the two RGBA values that are being compared.

The mix_colors() function no longer sets the alpha channel to 1, enabling
the user to choose semi-transparent colors if desired.

See also: https://gitlab.xfce.org/panel-plugins/xfce4-systemload-plugin/-/merge_requests/27#note_36436