Increase type-safety of the return types of callback functions
The impetus of this patch were the invalid return types of ButtonHandler,
DrawHandler and SizeChangeHandler.

The C++ implementation internally uses partial template specialization
which is needed to separate the type "void" from other data types.

The file gtk.cc contains comments with HTTP links to GTK and XFCE
documentation which can be used to verify the handler function signatures.