Initially hide run-in-terminal checkbox if the command field is empty (#23)
This patch modifies the visual appearance of the properties dialog so that
the checkboxes of "Associated command" are initially hidden if the
command text field is empty. Once the text field contains some characters,
the checkboxes remain visible until the dialog is closed, but are grayed
out (are logically disabled) if the text field becomes empty again.

The reason for the checkboxes to remain visible is that displaying them
enlarges the dialog window in the vertical direction, and so removing them
would either shrink the dialog window or show empty vertical space near the
bottom of the dialog window.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpugraph-plugin/-/issues/23